requirements.txt
openai
pillow
pyyaml  
Requests
selenium
tenacity
webdriver_manager
yaspin
bs4
anthropic